Move ResponseLogger.Lift to Logger, extract internal Logger type
